pyproject.toml
[project]
name = "eia-mcp"
version = "0.1.0"
description = "MCP server for EIA (Energy Information Administration) API - Electricity and Natural Gas data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
eia-mcp = "eia_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["eia_mcp"]
